reorganized example integrations
