+++ example/server/integrations/__init__.py
@@ -1,0 +1,19 @@
+from .deposit import MyDepositIntegration
+from .withdraw import MyWithdrawalIntegration
+from .sep31 import MySEP31ReceiverIntegration
+from .customers import MyCustomerIntegration
+from .info import info_integration
+from .fee import fee_integration
+from .quotes import MyQuoteIntegration
+from .rails import MyRailsIntegration
+
+__all__ = [
+    "MyDepositIntegration",
+    "MyWithdrawalIntegration",
+    "MyRailsIntegration",
+    "MyCustomerIntegration",
+    "MySEP31ReceiverIntegration",
+    "MyQuoteIntegration",
+    "info_integration",
+    "fee_integration",
+]
